// lib/types.ts
export interface NewsItem {
  headLineEnglish?: string;
  headLineNepali?: string;
  titleEnglish?: string;
  titleNepali?: string;
  title?: string;
  descriptionEnglish?: string;
  descriptionNepali?: string;
  newsSite?: string;
  author?: string;
  url?: string;
  thumbnail?: string;
  thumb?: string;
  topicA?: string;
  topicB?: string;
  topicC?: string;
  status?: string;
  autoGenerated?: boolean;
  createdAt?: number;
  snoozedUntil?: number;
  publishAt?: number;
}

export type NewsRecord = Record<string, NewsItem>;

export type NewsType = "pending" | "published";